Update the user model
Includes the bcrypt extension to use for setting the encrypted password
of the user.

Added getters and setters for password and password confirmation.

The setter and getter for password encrypts and checks the password of
the user provided respectively using bcrypt. Also added validations so
these do not have to be done by comparison in the controller.

This could be refined a bit more but that should be done with testing
:)